Stride & New Touch Handler
* Add new Touch Handler
* Add Support for non multiply 8 display (like qhd)
* Enhance Fling performance
* Add Precision for calibrating data. Now use float so it won't rounded into integer and more accurate in touch

+++ aroma-installer/src/aroma_ui.c
@@ -542,7 +542,7 @@ Value* AROMA_CALIBRATE(const char* name, State* state, int argc, Expr* argv[]) {
   if (ReadArgs(state, argv, 4, &dx, &ax, &dy, &ay) < 0){
       return NULL;
   }
-  atouch_set_calibrate(atoi(dx),atoi(ax),atoi(dy),atoi(ay));
+  atouch_set_calibrate((float) atof(dx),atoi(ax),(float) atof(dy),atoi(ay));
   free(dx);
   free(dy);
   free(ax);
